requirements.txt
certifi==2019.11.28
cffi==1.14.0
cycler==0.10.0
kiwisolver==1.1.0
matplotlib==3.1.3
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
numpy==1.18.1
olefile==0.46
Pillow==7.0.0
pycparser==2.20
pyparsing==2.4.6
python-dateutil==2.8.1
six==1.14.0
torch==1.4.0
torchvision==0.5.0
tornado==6.0.4
wincertstore==0.2
